CadastroFastAPI/requirements.txt
fastapi
uvicorn
requests
sqlalchemy
pydantic
email-validator
python-dotenv